Common ground for Aggragation and Computre
